test.sh: Make it work on fedora
Original test works on ubuntu for travis.  Modify so that we can run most
of the unit test on fedora.  For fedora when we test libStorageMgmt unit
test we test against installed package.

We are skipping the legacy unit test as it requires ZFS.

+++ test/test.sh
@@ -10,9 +10,12 @@ clean_up ()
 
     pkill lsmd || echo "Warning: Unable to end lsmd process"
 
-    python3-coverage report || echo "Warning: Unable to generate report"
+    $COV_CMD report || echo "Warning: Unable to generate report"
 
-    zpool destroy zfs_targetd || echo "Warning: unable to destroy zpool zfs_targetd"
+    if [ $FEDORA -eq 0 ]; then
+        zpool destroy zfs_targetd || echo "Warning: unable to destroy zpool zfs_targetd"
+    fi
+
     umount /mnt/btrfs || echo "Warning: , unable to unmount /mnt/btrfs"
     vgremove -f vg-targetd || echo "Warning: unable to remove VG vg-targetd"
     losetup -D || echo "Warning: Unable to remove all loopback devices"
@@ -21,6 +24,18 @@ clean_up ()
 
     exit $1
 }
+
+# We don't have ZFS in fedora for default repos.
+FEDORA=0
+if [ -f "/etc/redhat-release" ]; then
+    FEDORA=1
+fi
+
+if [ $FEDORA -eq 0 ]; then
+    COV_CMD=python3-coverage
+else
+    COV_CMD=coverage3
+fi
 
 # Handle a couple options to allow developers to be more productive in
 # their test environment.
@@ -46,10 +61,23 @@ if [  $TEARDOWN -eq 1 ]; then
 fi
 
 # Will use to see how well the test coverage is ...
-apt-get install python3-coverage
+if [ $FEDORA -eq 0 ]; then
+    apt-get install python3-coverage || clean_up 1
+else
+    # Travis yml file handles dependencies, we will do that
+    # here for other.
+    dnf install `cat test/fedora_pkgs.txt` -y -q || clean_up 1
+fi
 
 mkdir -p /etc/target || clean_up 1
 cp test/targetd.yaml /etc/target/ || clean_up 1
+
+# Remove zfs references from config file
+if [ $FEDORA -eq 1 ]; then
+    sed -e '/zfs_block_pools/ s/^#*/#/' -i /etc/target/targetd.yaml
+    sed -e '/fs_pools/ s/^#*/#/' -i /etc/target/targetd.yaml
+    echo "fs_pools: [/mnt/btrfs]" >> /etc/target/targetd.yaml
+fi
 
 # Needed for NFS functionality
 mkdir -p /etc/exports.d || clean_up 1
@@ -81,9 +109,11 @@ mkdir -p /mnt/btrfs || clean_up 1
 mount $loop2 /mnt/btrfs || clean_up 1
 
 # Create needed zfs
-zpool create zfs_targetd $loop3 || clean_up 1
-zfs create zfs_targetd/block_pool || clean_up 1
-zfs create zfs_targetd/fs_pool || clean_up 1
+if [ $FEDORA -eq 0 ]; then
+    zpool create zfs_targetd $loop3 || clean_up 1
+    zfs create zfs_targetd/block_pool || clean_up 1
+    zfs create zfs_targetd/fs_pool || clean_up 1
+fi
 
 # setup a sample nfs export file for parsing
 cp test/targetd_test_parse /tmp/. || clean_up 1
@@ -95,7 +125,7 @@ fi
 
 export PYTHONPATH=$(pwd)
 export TARGETD_NFS_EXPORT="/tmp/targetd_test_parse"
-python3-coverage run --omit=/usr/lib/python3/* scripts/targetd > /tmp/targetd.log 2>&1 &
+$COV_CMD run --omit=/usr/lib/python3/* scripts/targetd > /tmp/targetd.log 2>&1 &
 
 sleep 5 || clean_up 1
 echo "Dumping targetd output ..."
@@ -116,25 +146,27 @@ if [ $UNIT_ONLY -eq 1 ]; then
     clean_up 0
 fi
 
-# get/build/run libstoragemgmt tests
-TEST_DIR="$(pwd)/test"
-./test/lsm_test.sh $TEST_DIR
+
+# get/build/run or install libstoragemgmt tests
+./test/lsm_test.sh
+
 rc=$?
 if [ $rc -ne 0 ]; then
-  echo "Dumping targetd output on libstoragemgmt error ..."
-  cat /tmp/targetd.log
-  clean_up $rc
+    echo "Dumping targetd output on libstoragemgmt error ..."
+    cat /tmp/targetd.log
+    clean_up $rc
 fi
 
-# Run the actual legacy tests ...
-echo "Running client test ..."
-python3 client
-rc=$?
-if [ $rc -ne 0 ]; then
-    echo "Dumping targetd output on client error ..."
-    cat /tmp/targetd.log
-    clean_up
-    clean_up $rc
+# Run the actual legacy tests which only work with ZFS enabled
+if [ $FEDORA -eq 0 ]; then
+    echo "Running legacy client test ..."
+    python3 client
+    rc=$?
+    if [ $rc -ne 0 ]; then
+        echo "Dumping targetd output on client error ..."
+        cat /tmp/targetd.log
+        clean_up $rc
+    fi
 fi
 
 clean_up 0
